Removed gradient normalization/clipping
Previous commit utilized nn.utils.clip_grad_norm_ to attempt to solve an issue where training would result in NAN values. This appears to be related to bad hardware in my GPU. When running the code on the CPU it appears to be just fine. I have reverted to disabling this line to match my jupyter notebook, which ran just fine on the Udacity servers and gives further credence to the idea that this code works just fine, but that the errors are related to my personal machine and not the python code.